refactor(file-structure): adding data/firestore.ts for our db queries
